libdbg: Introduce kernel FSI driver backend
Creatively named 'kernel', this uses the SoftFSI aka OpenFSI aka fsi
kernel driver.

This backend triggers a probe of all attached FSI devices by the kernel.
It then This assumes a single slave appears in sysfs, and uses the 'raw'
device to perform reads and writes directly to the CFAM address space.

It has been successfully proved to not work on a Plametto:

 # ./pdbg -b kernel
 pdbg: Failed to read from 0x00000c09: Invalid argument

Tips are welcome. As are rants for how horribly I butchered the
implementation.

Signed-off-by: Joel Stanley <joel@jms.id.au>