libpdbg: Wrappers for cache inhibited adu_getmem and adu_putmem
Tacking a boolean parameter onto the end of the a function makes it hard
to read the code at call sites and know what is going on.

See Rusty's API levels, point 6 and 7:

 7. The obvious use is (probably) the correct one.
 6. The name tells you how to use it.

Cache inhibited is probably not what you're after, unless you know that
you're after it. These wrappers remove the boolean from the user
callable API.

Signed-off-by: Joel Stanley <joel@jms.id.au>