kernel: Support aspeed hardware FSI master
The paths to sysfs entries depend on the master driver used. This means
pdbg needs to use a different path when running against the hardware FSI
master present in the ast2600.

I do not intend for this patch to be merged as-is. We do need to use
this as a discussion point for a solution, probably in the kernel, that
fixes this.

Signed-off-by: Joel Stanley <joel@jms.id.au>